Complete test for 'pack' function factory
